requirements.txt
# runpod/pytorch:2.1.0-py3.10-cuda11.8.0-devel-ubuntu22.04

scipy==1.11.4
triton==2.0.0
altair==4.1.0
pandas==2.1.4
mediapipe==0.10.8

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

diffusers==0.18.0 
huggingface_hub==0.16.4

transformers==4.31.0
accelerate==0.21.0

controlnet-aux==0.0.7
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78